refactor(dashboard): consolidate 7 useMemo blocks into useDashboardSummary hook
Phase 3A implementation complete:
- Hook reuses existing domain types (PersonDaily, etc.)
- Direct copy-paste of original 7 useMemo blocks
- DashboardPage reduced by ~290 lines
- Zero logic changes (pure refactoring)
- typecheck ✅ build ✅ lint ✅
